refactor: changed file names + moved item view to its own file
